move cmd fixture to archiver test module
